+++ temba/tickets/migrations/0059_delete_old_exports.py
@@ -5,7 +5,7 @@ from django.core.files.storage import default_storage
 from django.db import migrations
 
 
-def delete_old_exports(apps, schema_editor):
+def delete_old_exports(apps, schema_editor):  # pragma: no cover
     ExportTicketsTask = apps.get_model("tickets", "ExportTicketsTask")
     num_deleted = 0
     num_skipped = 0
@@ -28,7 +28,7 @@ def delete_old_exports(apps, schema_editor):
         print(f"Deleted {num_deleted} old exports and skipped {num_skipped} old exports.")
 
 
-def reverse(apps, schema_editor):
+def reverse(apps, schema_editor):  # pragma: no cover
     pass