+++ temba/utils/export/models.py
@@ -1,150 +1,15 @@
 import gc
 import logging
-import os
-import time
 from datetime import datetime
 
-from smartmin.models import SmartModel
 from xlsxlite.writer import XLSXBook
 
-from django.core.files import File
 from django.core.files.temp import NamedTemporaryFile
-from django.db import models
 from django.http import HttpResponse
-from django.utils.translation import gettext_lazy as _
 
-from temba.assets.models import BaseAssetStore, get_asset_store
-from temba.utils import analytics
-from temba.utils.models import TembaUUIDMixin
 from temba.utils.text import clean_string
 
 logger = logging.getLogger(__name__)
-
-
-class BaseExportAssetStore(BaseAssetStore):
-    def is_asset_ready(self, asset):
-        return asset.status == BaseExport.STATUS_COMPLETE
-
-
-class BaseExport(TembaUUIDMixin, SmartModel):
-    """
-    Base class for export task models, i.e. contacts, messages and flow results
-    """
-
-    analytics_key = None
-    asset_type = None
-    notification_export_type = None
-
-    MAX_EXCEL_ROWS = 1_048_576
-    MAX_EXCEL_COLS = 16384
-
-    WIDTH_SMALL = 15
-    WIDTH_MEDIUM = 20
-    WIDTH_LARGE = 100
-
-    STATUS_PENDING = "P"
-    STATUS_PROCESSING = "O"
-    STATUS_COMPLETE = "C"
-    STATUS_FAILED = "F"
-    STATUS_CHOICES = (
-        (STATUS_PENDING, _("Pending")),
-        (STATUS_PROCESSING, _("Processing")),
-        (STATUS_COMPLETE, _("Complete")),
-        (STATUS_FAILED, _("Failed")),
-    )
-
-    # log progress after this number of exported objects have been exported
-    LOG_PROGRESS_PER_ROWS = 10000
-
-    org = models.ForeignKey("orgs.Org", on_delete=models.PROTECT, related_name="%(class)ss")
-    status = models.CharField(max_length=1, default=STATUS_PENDING, choices=STATUS_CHOICES)
-    num_records = models.IntegerField(null=True)
-
-    def perform(self):
-        """
-        Performs the actual export. If export generation throws an exception it's caught here and the task is marked
-        as failed.
-        """
-        from temba.notifications.types.builtin import ExportFinishedNotificationType
-
-        try:
-            self.update_status(self.STATUS_PROCESSING)
-
-            print(f"Started performing {self.analytics_key} with ID {self.id}")
-
-            start = time.time()
-
-            temp_file, extension, num_records = self.write_export()
-
-            get_asset_store(model=self.__class__).save(self.id, File(temp_file), extension)
-
-            # remove temporary file
-            if hasattr(temp_file, "delete"):
-                if temp_file.delete is False:  # pragma: no cover
-                    os.unlink(temp_file.name)
-            else:  # pragma: no cover
-                os.unlink(temp_file.name)
-
-        except Exception as e:  # pragma: no cover
-            logger.error(f"Unable to perform export: {str(e)}", exc_info=True)
-            self.update_status(self.STATUS_FAILED)
-            print(f"Failed to complete {self.analytics_key} with ID {self.id}")
-
-            raise e  # log the error to sentry
-        else:
-            self.update_status(self.STATUS_COMPLETE, num_records)
-            elapsed = time.time() - start
-            print(f"Completed {self.analytics_key} with ID {self.id} in {elapsed:.1f} seconds")
-            analytics.track(self.created_by, "temba.%s_latency" % self.analytics_key, properties=dict(value=elapsed))
-
-            ExportFinishedNotificationType.create(self)
-        finally:
-            gc.collect()  # force garbage collection
-
-    def write_export(self):  # pragma: no cover
-        """
-        Should return 1) file handle for a temporary file, 2) file extension, 3) count of items exported
-        """
-        pass
-
-    def update_status(self, status: str, num_records: int = None):
-        self.status = status
-        self.num_records = num_records
-        self.save(update_fields=("status", "num_records", "modified_on"))
-
-    def get_download_url(self) -> str:
-        asset_store = get_asset_store(model=self.__class__)
-        return asset_store.get_asset_url(self.id)
-
-    def get_notification_scope(self) -> str:
-        return f"{self.notification_export_type}:{self.id}"
-
-    class Meta:
-        abstract = True
-
-
-class BaseDateRangeExport(BaseExport):
-    """
-    Base export class for exports that have a date range.
-    """
-
-    start_date = models.DateField()
-    end_date = models.DateField()
-
-    class Meta:
-        abstract = True
-
-
-class BaseItemWithContactExport(BaseDateRangeExport):
-    """
-    Base export class for exports that are an item with an associated contact.
-    """
-
-    with_fields = models.ManyToManyField("contacts.ContactField", related_name="%(class)s_exports")
-    with_groups = models.ManyToManyField("contacts.ContactGroup", related_name="%(class)s_exports")
-
-    class Meta:
-        abstract = True
 
 
 def prepare_value(value, tz=None):
@@ -170,6 +35,9 @@ class MultiSheetExporter:
     Utility to aid writing a stream of rows which may exceed the 1048576 limit on rows per sheet, and require adding
     new sheets.
     """
+
+    MAX_EXCEL_ROWS = 1_048_576
+    MAX_EXCEL_COLS = 16384
 
     def __init__(self, base_sheet_name: str, headers: list, tz):
         self.base_sheet_name = base_sheet_name
@@ -199,7 +67,7 @@ class MultiSheetExporter:
         assert len(values) == len(self.headers), "need same number of column values as column headers"
 
         # time for a new sheet? do it
-        if self.sheet_row > BaseExport.MAX_EXCEL_ROWS:
+        if self.sheet_row > self.MAX_EXCEL_ROWS:
             self._add_sheet()
 
         self.sheet.append_row(*[prepare_value(v, self.tz) for v in values])